app/requirements.txt
python-decouple
aiohttp
aioretry
sqlalchemy
psycopg2
requests
web3